Fix parsing media URLs from submissions
For images check for "i.redd.it" in "domain" OR "image" in "post_hint".
Post hints aren't present in all subreddits.
For videos check for "v.redd.it" in "domain" AND whether "is_video" is true.
Submissions from domain "v.redd.it", but without "is_video" don't have necessary URLs.